simplify finder, run cargo fmt
