+++ state-rpc/src/main.rs
@@ -1,11 +1,11 @@
-use std::path::Path;
 use std::collections::HashMap;
+use std::path::Path;
 
 use derive_more::Display;
 
-use evm_state::*;
-use evm_state::rand::Rng;
 use evm_rpc::FormatHex;
+use evm_state::rand::Rng;
+use evm_state::*;
 
 mod finder;
 
@@ -21,14 +21,11 @@ struct Rpc {}
 
 #[tonic::async_trait]
 impl Backend for Rpc {
-    async fn ping(
-        &self,
-        request: Request<()>,
-    ) -> Result<Response<PingReply>, Status> {
+    async fn ping(&self, request: Request<()>) -> Result<Response<PingReply>, Status> {
         println!("Got a request: {:?}", request);
 
         let reply = app_grpc::PingReply {
-            message: "Ok".to_string()
+            message: "Ok".to_string(),
         };
 
         Ok(Response::new(reply))
@@ -42,16 +39,13 @@ impl Backend for Rpc {
 
         let stringified_key = request.into_inner().hash;
 
-        let state_root_str = std::fs::read_to_string("./.tmp/state_root.txt").expect("get the state root");
         let dir = Path::new("./.tmp/db/");
-
-        let state_root = H256::from_hex(&state_root_str).expect("get hash from &str");
 
         let db_handle = Storage::open_persistent(dir, true).expect("could not open database");
 
         let key = H256::from_hex(&stringified_key).expect("get hash from &str");
 
-        let finder = finder::Finder::new(db_handle, key);
+        let finder = finder::Finder::new(db_handle);
         let maybe_bytes = finder.find(key);
 
         let response = if let Ok(Some(bytes)) = maybe_bytes {
@@ -60,9 +54,7 @@ impl Backend for Rpc {
             HashMap::new()
         };
 
-        let reply = app_grpc::BlockData {
-            data: response
-        };
+        let reply = app_grpc::BlockData { data: response };
 
         Ok(Response::new(reply))
     }
@@ -71,32 +63,24 @@ impl Backend for Rpc {
         &self,
         request: Request<app_grpc::MultiHash>,
     ) -> Result<Response<app_grpc::MultiBlockData>, Status> {
-
         let stringified_keys = request.into_inner().hashes;
-
         let mut response = HashMap::new();
-
-        stringified_keys.into_iter().for_each(|stringified_key| {
-let state_root_str = std::fs::read_to_string("./.tmp/state_root.txt").expect("get the state root");
         let dir = Path::new("./.tmp/db/");
 
-        let state_root = H256::from_hex(&state_root_str).expect("get hash from &str");
+        stringified_keys.into_iter().for_each(|stringified_key| {
+            let db_handle = Storage::open_persistent(dir, true).expect("could not open database");
 
-        let db_handle = Storage::open_persistent(dir, true).expect("could not open database");
+            let key = H256::from_hex(&stringified_key).expect("get hash from &str");
 
-        let key = H256::from_hex(&stringified_key).expect("get hash from &str");
-
-        let finder = finder::Finder::new(db_handle, key);
-        let maybe_bytes = finder.find(key);
+            let finder = finder::Finder::new(db_handle);
+            let maybe_bytes = finder.find(key);
 
             if let Ok(Some(bytes)) = maybe_bytes {
                 response.insert(stringified_key, bytes);
             }
         });
 
-        let reply = app_grpc::MultiBlockData {
-            data: response
-        };
+        let reply = app_grpc::MultiBlockData { data: response };
 
         Ok(Response::new(reply))
     }
@@ -104,14 +88,14 @@ let state_root_str = std::fs::read_to_string("./.tmp/state_root.txt").expect("ge
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    // For the future state movement
+    // let state_root_str = std::fs::read_to_string("./.tmp/state_root.txt").expect("get the state root");
+    // let state_root = H256::from_hex(&state_root_str).expect("get hash from &str");
+
     let addr = "127.0.0.1:8000".parse()?;
     let backend = BackendServer::new(Rpc {});
 
-    Server::builder()
-       .add_service(backend)
-       .serve(addr)
-       .await?;
+    Server::builder().add_service(backend).serve(addr).await?;
 
     Ok(())
 }
-
